+++ shoop/core/api/orders.py
@@ -6,26 +6,38 @@
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
 from django.utils.timezone import now
-from rest_framework.serializers import ModelSerializer
+from rest_framework import serializers, status
+from rest_framework.decorators import detail_route
+from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
 
-from shoop.core.models import MutableAddress, Order, OrderLine, OrderStatus
+from shoop.core.models import (
+    MutableAddress, Order, OrderLine, OrderStatus, Payment
+)
+from shoop.utils.money import Money
 
 
-class OrderLineSerializer(ModelSerializer):
+class OrderLineSerializer(serializers.ModelSerializer):
     class Meta:
         model = OrderLine
 
 
-class AddressSerializer(ModelSerializer):
+class AddressSerializer(serializers.ModelSerializer):
     class Meta:
         model = MutableAddress
 
 
-class OrderSerializer(ModelSerializer):
+class PaymentSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Payment
+        fields = ("payment_identifier", "amount_value", "description")
+
+
+class OrderSerializer(serializers.ModelSerializer):
     lines = OrderLineSerializer(many=True, read_only=True)
     billing_address = AddressSerializer(read_only=True)
     shipping_address = AddressSerializer(read_only=True)
+    payments = PaymentSerializer(many=True, read_only=True)
 
     def get_fields(self):
         fields = super(OrderSerializer, self).get_fields()
@@ -48,3 +60,31 @@ class OrderSerializer(ModelSerializer):
 class OrderViewSet(ModelViewSet):
     serializer_class = OrderSerializer
     queryset = Order.objects.all()
+
+    @detail_route(methods=['post'])
+    def create_payment(self, request, pk=None):
+        return _handle_payment_creation(request, self.get_object())
+
+    @detail_route(methods=['post'])
+    def set_fully_paid(self, request, pk=None):
+        order = self.get_object()
+        if order.is_paid():
+            return Response({"status": "order is already fully paid"})
+
+        request.data["currency"] = order.currency
+        request.data["amount_value"] = (order.taxful_total_price_value - order.get_total_paid_amount().value)
+        return _handle_payment_creation(request, order)
+
+
+def _handle_payment_creation(request, order):
+    serializer = PaymentSerializer(data=request.data)
+    if serializer.is_valid():
+        data = serializer.validated_data
+        order.create_payment(
+            Money(data["amount_value"], order.currency),
+            data["payment_identifier"],
+            data.get("description", "")
+        )
+        return Response({'status': 'payment created'}, status=status.HTTP_201_CREATED)
+    else:
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)